Update hodor_app.js
Moved keypad pin details to the configuration file, as well as adding the ability to specify timeout period between key depressions. Also added support for a sensor pin so that a magnetic switch could be installed so that the app could keep track of open/closed door state. Also added ability for the "C" key to close the door if it's open and a valid code was used within the last 90 seconds (which is configurable through the config file).