Added ability to generate temporary codes
If someone attempts to gain entry outside of their normally allowed windows of access a temporary code is generated on their behalf and sent to the paging/texting group associated with their entry (not the person represented by the entry). It will expire after config.temp_code_ttl_minutes minutes. This feature allows the site administrator (the person receiving the text) to decide whether to share the temporary code with the person seeking entry.